+++ CHANGES.rst
@@ -11,6 +11,10 @@ Changelog of dask-geomodeling
 - The MemorySource and elementwise blocks now return None for 'extent' and
   'geometry' if they are empty.
 
+- Preserve functionality of the geometry.Difference block with geopandas 0.6.
+  When taking the difference of a geometry with a missing geometry (A - None),
+  geopandas < 0.6 returned A as result, while >= 0.6 returns None as result.
+
 
 2.0.3 (2019-10-08)
 ------------------