+++ CHANGES.rst
@@ -15,6 +15,9 @@ Changelog of dask-geomodeling
   When taking the difference of a geometry with a missing geometry (A - None),
   geopandas < 0.6 returned A as result, while >= 0.6 returns None as result.
 
+
+- Implemented the 'columns' attribute for GeometryFileSource.
+
 - Fixed the projection attribute of elementwise raster blocks in case one of
   the arguments is a number and not a Block instance.