+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog of dask-geomodeling
 ===================================================
 
 
-2.0.4 (unreleased)
+2.0.4 (2019-11-01)
 ------------------
 
 - Fixed propagation of the 'extent' and 'geometry' attributes through the